+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/one_tap/slider/ConsumerCreditsFragment.java
@@ -1,5 +1,6 @@
 package com.mercadopago.android.px.internal.features.one_tap.slider;
 
+import android.graphics.Color;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -7,26 +8,32 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.widget.AppCompatTextView;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.fragment.app.Fragment;
 import com.meli.android.carddrawer.model.CardDrawerView;
 import com.mercadopago.android.px.R;
+import com.mercadopago.android.px.core.presentation.extensions.DrawableExtKt;
 import com.mercadopago.android.px.internal.di.MapperProvider;
 import com.mercadopago.android.px.internal.features.payment_result.remedies.RemediesLinkableMapper;
+import com.mercadopago.android.px.internal.font.FontHelper;
+import com.mercadopago.android.px.internal.font.PxFont;
 import com.mercadopago.android.px.internal.util.ViewUtils;
 import com.mercadopago.android.px.internal.view.LinkableTextView;
 import com.mercadopago.android.px.internal.viewmodel.DisableConfiguration;
 import com.mercadopago.android.px.internal.viewmodel.drawables.ConsumerCreditsDrawableFragmentItem;
 import com.mercadopago.android.px.model.ConsumerCreditsDisplayInfo;
+import com.mercadopago.android.px.model.internal.Text;
 
 public class ConsumerCreditsFragment extends PaymentMethodFragment<ConsumerCreditsDrawableFragmentItem> {
 
-    private ConstraintLayout creditsLagout;
+    private ConstraintLayout creditsLayout;
     private ImageView background;
     private ImageView logo;
     private LinkableTextView topText;
     private LinkableTextView bottomText;
+    private ViewGroup tagContainer;
     private final RemediesLinkableMapper remediesLinkableMapper = MapperProvider.INSTANCE.getRemediesLinkableMapper();
 
     @NonNull
@@ -46,16 +53,29 @@ public class ConsumerCreditsFragment extends PaymentMethodFragment<ConsumerCredi
     @Override
     public void initializeViews(@NonNull final View view) {
         super.initializeViews(view);
-        creditsLagout = view.findViewById(R.id.credits_layout);
+        creditsLayout = view.findViewById(R.id.credits_layout);
         background = view.findViewById(R.id.background);
         logo = view.findViewById(R.id.logo);
         topText = view.findViewById(R.id.top_text);
         bottomText = view.findViewById(R.id.bottom_text);
+        tagContainer = view.findViewById(R.id.card_tag_container);
         final ConsumerCreditsDisplayInfo displayInfo = model.metadata.displayInfo;
         tintBackground(background, displayInfo.color);
         showDisplayInfo(displayInfo);
+        if (model.tag != null) {
+            showTag(model.tag);
+        }
         configureListener();
         view.setImportantForAccessibility(View.IMPORTANT_FOR_ACCESSIBILITY_NO_HIDE_DESCENDANTS);
+    }
+
+    private void showTag(@NonNull final Text tag) {
+        AppCompatTextView text = tagContainer.findViewById(R.id.card_cc_tag);
+        tagContainer.setVisibility(View.VISIBLE);
+        text.setText(tag.getMessage());
+        ViewUtils.setTextColor(text, tag.getTextColor());
+        DrawableExtKt.setColorFilter(text.getBackground(),tag.getBackgroundColor(),Color.WHITE);
+        FontHelper.setFont(text, PxFont.from(tag.getWeight()));
     }
 
     @Override
@@ -91,7 +111,7 @@ public class ConsumerCreditsFragment extends PaymentMethodFragment<ConsumerCredi
     public void disable() {
         super.disable();
         final DisableConfiguration disableConfiguration = new DisableConfiguration(getContext());
-        ViewUtils.grayScaleViewGroup(creditsLagout);
+        ViewUtils.grayScaleViewGroup(creditsLayout);
         background.clearColorFilter();
         background.setImageResource(0);
         background.setBackgroundColor(disableConfiguration.getBackgroundColor());
@@ -104,12 +124,12 @@ public class ConsumerCreditsFragment extends PaymentMethodFragment<ConsumerCredi
 
     private void centerLogo() {
         final ConstraintSet constraintSet = new ConstraintSet();
-        constraintSet.clone(creditsLagout);
-        constraintSet.connect(logo.getId(), ConstraintSet.LEFT, creditsLagout.getId(), ConstraintSet.LEFT, 0);
-        constraintSet.connect(logo.getId(), ConstraintSet.RIGHT, creditsLagout.getId(), ConstraintSet.RIGHT, 0);
-        constraintSet.connect(logo.getId(), ConstraintSet.TOP, creditsLagout.getId(), ConstraintSet.TOP, 0);
-        constraintSet.connect(logo.getId(), ConstraintSet.BOTTOM, creditsLagout.getId(), ConstraintSet.BOTTOM, 0);
-        constraintSet.applyTo(creditsLagout);
+        constraintSet.clone(creditsLayout);
+        constraintSet.connect(logo.getId(), ConstraintSet.LEFT, creditsLayout.getId(), ConstraintSet.LEFT, 0);
+        constraintSet.connect(logo.getId(), ConstraintSet.RIGHT, creditsLayout.getId(), ConstraintSet.RIGHT, 0);
+        constraintSet.connect(logo.getId(), ConstraintSet.TOP, creditsLayout.getId(), ConstraintSet.TOP, 0);
+        constraintSet.connect(logo.getId(), ConstraintSet.BOTTOM, creditsLayout.getId(), ConstraintSet.BOTTOM, 0);
+        constraintSet.applyTo(creditsLayout);
     }
 
     @Override