+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+_XX_04_2022
+* FEATURE - New layout consumer credits and new tag.
+
 ## VERSION 4.110.0
 _07_04_2022_
 * FIX - Crash when installments of payerCosts for credit cards is null.